+++ runtime/src/ujima/linux/system.clj
@@ -2,7 +2,7 @@
   (:require [clojure.string :as str]
             [lib.io :refer [slurp-text]]
             [lib.shell :refer [$? $!]]
-            [ujima.linux.sudo :refer [sudo$!]]))
+            [ujima.linux.sudo :refer [sudo$! sudo$?]]))
 
 
 (defn hostname []
@@ -27,6 +27,13 @@
 
 (defn uptime-minutes []
   (some-> (slurp-text "/proc/uptime" nil) (str/split #"\s") first parse-double (/ 60) long))
+
+
+(defn clock!
+  "Set the wall clock to EPOCH-MS; the RTC write is best-effort (no RTC, no error)."
+  [epoch-ms]
+  (sudo$! date -u -s [(str "@" (quot epoch-ms 1000))])
+  (sudo$? hwclock -w))
 
 
 (defn reboot! []